refactor: move commont types to typescript
