+++ packages/express/index.ts
@@ -6,7 +6,7 @@ export { default as asyncHandler } from './async-handler.js'
 export { default as errorHandler } from './error-handler.js'
 export * from './events-log.js'
 export * from './find-utils.js'
-export * from '../http-errors.js'
+export * from '@data-fair/lib-utils/http-errors.js'
 export * from './req-origin.js'
 export * from './site.js'
 export * from './serve-spa.js'
