+++ packages/vuetify/package.json
@@ -3,9 +3,10 @@
   "version": "0.0.0",
   "description": "Components based on the Vuetify 3 UI lib for the data-fair stack.",
   "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
+  "files": [
+    "**/*.js",
+    "**/*.d.ts"
+  ],
   "author": "",
   "license": "MIT",
   "peerDependencies": {
